change data type

+++ src/main/java/swp391/aistudyhub/dto/projection/StorageUsageResponse.java
@@ -1,5 +1,5 @@
 package swp391.aistudyhub.dto.projection;
 
 public interface StorageUsageResponse {
-    Long getUsedQuota();
+    Double getUsedQuota();
 }
