+++ src/main/java/swp391/aistudyhub/dto/request/SystemConfigDTO.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class SystemConfigDTO {
     private Integer maxDailyChatTokens;
-    private Long totalStorageQuotaGb;
-    private Long maxFileSizeMb;
+    private Double totalStorageQuotaGb;
+    private Double maxFileSizeMb;
     private String allowedFileTypes;
 
 }
